requirements.txt
requests
aiohttp
pillow
